requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_kmp_llvm
alsa-lib=1.2.8=h166bdaf_0
aom=3.3.0=h27087fc_1
appdirs=1.4.4=pypi_0
attr=2.5.1=h166bdaf_1
blas=2.116=openblas
blas-devel=3.9.0=16_linux64_openblas
brotli=1.0.9=h166bdaf_8
brotli-bin=1.0.9=h166bdaf_8
brotlipy=0.7.0=py310h5764c6d_1005
bzip2=1.0.8=h7f98852_4
ca-certificates=2022.12.7=ha878542_0
cairo=1.16.0=ha61ee94_1014
certifi=2022.12.7=pyhd8ed1ab_0
cffi=1.15.1=py310h255011f_3
charset-normalizer=2.1.1=pyhd8ed1ab_0
click=8.1.3=pypi_0
contourpy=1.0.7=py310hdf3cbec_0
cpuonly=2.0=0
cryptography=39.0.0=py310h34c0648_0
cycler=0.11.0=pyhd8ed1ab_0
dbus=1.13.6=h5008d03_3
docker-pycreds=0.4.0=pypi_0
expat=2.5.0=h27087fc_0
ffmpeg=5.0.1=h964e5f1_2
fftw=3.3.10=nompi_hf0379b8_106
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=hab24e00_0
fontconfig=2.14.2=h14ed4e7_0
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
fonttools=4.38.0=py310h5764c6d_1
freetype=2.12.1=hca18f0e_1
gettext=0.21.1=h27087fc_0
gitdb=4.0.10=pypi_0
gitpython=3.1.31=pypi_0
glib=2.74.1=h6239696_1
glib-tools=2.74.1=h6239696_1
gmp=6.2.1=h58526e2_0
gnutls=3.6.13=h85f3911_1
graphite2=1.3.13=h58526e2_1001
gst-plugins-base=1.22.0=h4243ec0_0
gstreamer=1.22.0=h25f0c4b_0
gstreamer-orc=0.4.33=h166bdaf_0
harfbuzz=6.0.0=h8e241bc_0
icu=70.1=h27087fc_0
idna=3.4=pyhd8ed1ab_0
jack=1.9.22=h11f4161_0
jpeg=9e=h166bdaf_2
keyutils=1.6.1=h166bdaf_0
kiwisolver=1.4.4=py310hbf28c38_1
krb5=1.20.1=h81ceb04_0
lame=3.100=h166bdaf_1003
lcms2=2.14=hfd0df8a_1
ld_impl_linux-64=2.40=h41732ed_0
lerc=4.0.0=h27087fc_0
libblas=3.9.0=16_linux64_openblas
libbrotlicommon=1.0.9=h166bdaf_8
libbrotlidec=1.0.9=h166bdaf_8
libbrotlienc=1.0.9=h166bdaf_8
libcap=2.66=ha37c62d_0
libcblas=3.9.0=16_linux64_openblas
libclang=15.0.7=default_had23c3d_1
libclang13=15.0.7=default_h3e3d535_1
libcups=2.3.3=h36d4200_3
libdb=6.2.32=h9c3ff4c_0
libdeflate=1.17=h0b41bf4_0
libdrm=2.4.114=h166bdaf_0
libedit=3.1.20191231=he28a2e2_2
libevent=2.1.10=h28343ad_4
libffi=3.4.2=h7f98852_5
libflac=1.4.2=h27087fc_0
libgcc-ng=12.2.0=h65d4601_19
libgcrypt=1.10.1=h166bdaf_0
libgfortran-ng=12.2.0=h69a702a_19
libgfortran5=12.2.0=h337968e_19
libglib=2.74.1=h606061b_1
libgomp=12.2.0=h65d4601_19
libgpg-error=1.46=h620e276_0
libhwloc=2.8.0=h32351e8_1
libiconv=1.17=h166bdaf_0
libjpeg-turbo=2.1.4=h166bdaf_0
liblapack=3.9.0=16_linux64_openblas
liblapacke=3.9.0=16_linux64_openblas
libllvm15=15.0.7=hadd5161_0
libnsl=2.0.0=h7f98852_0
libogg=1.3.4=h7f98852_1
libopenblas=0.3.21=pthreads_h78a6416_3
libopus=1.3.1=h7f98852_1
libpciaccess=0.17=h166bdaf_0
libpng=1.6.39=h753d276_0
libpq=15.2=hb675445_0
libprotobuf=3.21.12=h3eb15da_0
libsndfile=1.2.0=hb75c966_0
libsqlite=3.40.0=h753d276_0
libstdcxx-ng=12.2.0=h46fd767_19
libsystemd0=252=h2a991cd_0
libtiff=4.5.0=h6adf6a1_2
libtool=2.4.7=h27087fc_0
libudev1=253=h0b41bf4_0
libuuid=2.32.1=h7f98852_1000
libva=2.17.0=h0b41bf4_0
libvorbis=1.3.7=h9c3ff4c_0
libvpx=1.11.0=h9c3ff4c_3
libwebp-base=1.2.4=h166bdaf_0
libxcb=1.13=h7f98852_1004
libxkbcommon=1.5.0=h79f4944_0
libxml2=2.10.3=h7463322_0
libzlib=1.2.13=h166bdaf_4
llvm-openmp=15.0.7=h0cdce71_0
lz4-c=1.9.4=hcb278e6_0
matplotlib=3.7.0=py310hff52083_0
matplotlib-base=3.7.0=py310he60537e_0
mkl=2022.2.1=h84fe81f_16997
mkl-devel=2022.2.1=ha770c72_16998
mkl-include=2022.2.1=h84fe81f_16997
mpg123=1.31.2=hcb278e6_0
munkres=1.1.4=pyh9f0ad1d_0
mysql-common=8.0.32=ha901b37_0
mysql-libs=8.0.32=hd7da12d_0
ncurses=6.3=h27087fc_1
nettle=3.6=he412f7d_0
networkx=3.0=pypi_0
ninja=1.11.0=h924138e_0
nspr=4.35=h27087fc_0
nss=3.88=he45b914_0
numpy=1.24.1=py310h08bbf29_0
openblas=0.3.21=pthreads_h320a7e8_3
openh264=2.1.1=h780b84a_0
openjpeg=2.5.0=hfec8fc6_2
openssl=3.0.8=h0b41bf4_0
packaging=23.0=pyhd8ed1ab_0
pathtools=0.1.2=pypi_0
pcre2=10.40=hc3806b6_0
pillow=9.4.0=py310h4927cde_0
pip=23.0=pyhd8ed1ab_0
pixman=0.40.0=h36c2ea0_0
ply=3.11=py_1
protobuf=4.22.0=pypi_0
pthread-stubs=0.4=h36c2ea0_1001
pulseaudio=16.1=ha8d29e2_1
pycparser=2.21=pyhd8ed1ab_0
pyopenssl=23.0.0=pyhd8ed1ab_0
pyparsing=3.0.9=pyhd8ed1ab_0
pyqt=5.15.7=py310hab646b1_3
pyqt5-sip=12.11.0=py310heca2aa9_3
pysocks=1.7.1=pyha2e5f31_6
python=3.10.8=h4a9ceb5_0_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.10=3_cp310
pytorch=1.13.1=cpu_py310hd11e9c7_0
pytorch-mutex=1.0=cpu
pyyaml=6.0=pypi_0
qt-main=5.15.8=h5d23da1_6
readline=8.1.2=h0f457ee_0
requests=2.28.2=pyhd8ed1ab_0
sentry-sdk=1.16.0=pypi_0
setproctitle=1.3.2=pypi_0
setuptools=66.1.1=pyhd8ed1ab_0
sip=6.7.7=py310heca2aa9_0
six=1.16.0=pyh6c4a22f_0
sleef=3.5.1=h9b69904_2
smmap=5.0.0=pypi_0
svt-av1=1.1.0=h27087fc_1
tbb=2021.7.0=h924138e_1
tk=8.6.12=h27826a3_0
toml=0.10.2=pyhd8ed1ab_0
torchaudio=0.13.1=py310_cpu
torchsummary=1.5.1=pypi_0
torchvision=0.14.1=cpu_py310hdfb2906_0
tornado=6.2=py310h5764c6d_1
typing_extensions=4.4.0=pyha770c72_0
tzdata=2022g=h191b570_0
unicodedata2=15.0.0=py310h5764c6d_0
urllib3=1.26.14=pyhd8ed1ab_0
wandb=0.13.10=pypi_0
wheel=0.38.4=pyhd8ed1ab_0
x264=1!161.3030=h7f98852_1
x265=3.5=h924138e_3
xcb-util=0.4.0=h166bdaf_0
xcb-util-image=0.4.0=h166bdaf_0
xcb-util-keysyms=0.4.0=h166bdaf_0
xcb-util-renderutil=0.3.9=h166bdaf_0
xcb-util-wm=0.4.1=h166bdaf_0
xorg-fixesproto=5.0=h7f98852_1002
xorg-kbproto=1.0.7=h7f98852_1002
xorg-libice=1.0.10=h7f98852_0
xorg-libsm=1.2.3=hd9c2040_1000
xorg-libx11=1.7.2=h7f98852_0
xorg-libxau=1.0.9=h7f98852_0
xorg-libxdmcp=1.1.3=h7f98852_0
xorg-libxext=1.3.4=h7f98852_1
xorg-libxfixes=5.0.3=h7f98852_1004
xorg-libxrender=0.9.10=h7f98852_1003
xorg-renderproto=0.11.1=h7f98852_1002
xorg-xextproto=7.3.0=h7f98852_1002
xorg-xproto=7.0.31=h7f98852_1007
xz=5.2.6=h166bdaf_0
zlib=1.2.13=h166bdaf_4
zstd=1.5.2=h3eb15da_6
